Add empty state to Front-end playlist list
